Fix transport cc large negative delta unmarshal
Large deltas are signed. But unmarshal method doesn't correctly handle
them returning a bigger positive number due to direct casting to int64.

This patch fixes this issue and adds a related test.

+++ transport_layer_cc.go
@@ -273,7 +273,7 @@ func (r *RecvDelta) Unmarshal(rawPacket []byte) error {
 	}
 
 	r.Type = TypeTCCPacketReceivedLargeDelta
-	r.Delta = TypeTCCDeltaScaleFactor * int64(binary.BigEndian.Uint16(rawPacket))
+	r.Delta = TypeTCCDeltaScaleFactor * int64(int16(binary.BigEndian.Uint16(rawPacket)))
 	return nil
 }
